* Add tracer to all test types * In message converter ignores unknown property
