+++ rqueue-spring-boot-starter/src/test/java/com/github/sonus21/rqueue/spring/boot/tests/SpringBootUnitTest.java
@@ -16,6 +16,7 @@
 
 package com.github.sonus21.rqueue.spring.boot.tests;
 
+import com.github.sonus21.junit.TestTracerExtension;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -29,5 +30,5 @@ import org.mockito.junit.jupiter.MockitoExtension;
 @Tag("springBoot")
 @Tag("unit")
 @Tag("springBootUnit")
-@ExtendWith({MockitoExtension.class})
+@ExtendWith({MockitoExtension.class, TestTracerExtension.class})
 public @interface SpringBootUnitTest {}
