+++ rqueue-spring/src/test/java/com/github/sonus21/rqueue/spring/tests/unit/RqueueMessageConfigTest.java
@@ -1,17 +1,17 @@
 /*
- * Copyright 2020 Sonu Kumar
+ *  Copyright 2021 Sonu Kumar
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
- *       https://www.apache.org/licenses/LICENSE-2.0
+ *         https://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and limitations under the License.
+ *
  */
 
 package com.github.sonus21.rqueue.spring.tests.unit;
@@ -21,6 +21,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 
+import com.github.sonus21.TestBase;
 import com.github.sonus21.rqueue.config.SimpleRqueueListenerContainerFactory;
 import com.github.sonus21.rqueue.converter.GenericMessageConverter;
 import com.github.sonus21.rqueue.core.DefaultRqueueMessageConverter;
@@ -43,12 +44,17 @@ import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.messaging.converter.MessageConverter;
 
 @SpringUnitTest
-class RqueueMessageConfigTest {
-  @Mock private SimpleRqueueListenerContainerFactory simpleRqueueListenerContainerFactory;
-  @Mock private BeanFactory beanFactory;
-  @Mock private RqueueMessageTemplate rqueueMessageTemplate;
-  @InjectMocks private RqueueListenerConfig rqueueMessageConfig;
-  private List<MessageConverter> messageConverterList = new ArrayList<>();
+class RqueueMessageConfigTest extends TestBase {
+
+  @Mock
+  private SimpleRqueueListenerContainerFactory simpleRqueueListenerContainerFactory;
+  @Mock
+  private BeanFactory beanFactory;
+  @Mock
+  private RqueueMessageTemplate rqueueMessageTemplate;
+  @InjectMocks
+  private RqueueListenerConfig rqueueMessageConfig;
+  private final List<MessageConverter> messageConverterList = new ArrayList<>();
 
   @BeforeEach
   public void init() {
